whole gui in the middle of refactioring to use grid panes instead of absolute positioning

+++ src/core/Coordinates.java
@@ -69,6 +69,6 @@ public class Coordinates {
   }
 
   public String toString() {
-    return "(" + x + ", " + y + ")";
+    return "(" + (int) x + ", " + (int) y + ")";
   }
 }
